Added link from retrieval to confluence
