+++ hubble-core/src/main/scala/team/supernova/cassandra/ClusterEnv.scala
@@ -1,12 +1,11 @@
 package team.supernova.cassandra
 
-import team.supernova.graphite.GraphiteMetric
+import team.supernova.HubbleApp.GraphiteConfig
 
 case class ClusterEnv( cluster_name: String,
                        graphana: String,
-                       graphite_template: String,
+                       graphiteConfig: GraphiteConfig,
                        graphite: Map[String, String],
-                       graphite_metrics: List[GraphiteMetric],
                        hosts: Array[String],
                        ops_pword: String,
                        ops_uname: String,
