+++ hubble-core/src/test/scala/team/supernova/graphite/AuthorizedMetricRetrieverSpec.scala
@@ -15,12 +15,19 @@ with GraphiteFixture{
   val clusterInstance: ClusterEnv = cassandragroup.head.envs.last
 
   describe("graphite metric retrieval"){
+
     it  ("should retrieve metric value") {
-      val metric = clusterInstance.graphite_metrics.head
+      val metric = clusterInstance.graphiteConfig.graphite_metrics.head
       val measure = new AuthorizedMetricRetriever(metric.url_template, metric.func, graphiteUserName, graphitePassword)
         .retrieve(clusterInstance.graphite)
       measure should not be None
       measure.get should be > 0D
+    }
+
+    it  ("should retrieve all metric values") {
+      val measures = AuthorizedMetricRetriever.retrieveAll(clusterInstance)
+      measures.size should be > 1
+      measures.filter(_._2.isEmpty) should have size 0
     }
 
   }
